Forward SIGTERM on UNIX.
SIGINT is generally sent to all processes connected to the virtual
terminal when typing Ctrl-C. SIGTERM is generally sent to a single
process. In that case, it is wise not to ignore that signal but to
forward it to the child process instead to give it the opportunity
to handle it.